+++ components/Buttons/SlideSelect/Options.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Children } from '@components/types';
+import { Children } from 'types';
 
 export const MatchingEngine: React.FC<{ title: string; subTitle: string }> = ({ title, subTitle }) => (
     <div className="m-auto text-center">
