+++ components/SummaryInfo/AccountMarginInfo.tsx
@@ -7,8 +7,8 @@ import {
     calcTotalMargin,
     calcMinimumMargin,
 } from '@tracer-protocol/tracer-utils';
-import { Section } from './';
-import { UserBalance } from '@components/types';
+import { Section } from '@components/General';
+import { UserBalance } from 'types';
 
 interface IProps {
     balance: UserBalance | undefined;
