+++ components/SummaryInfo/MarketInfo.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { Section, FundingRate } from './';
+import { Section } from '@components/General';
+import { FundingRateGraphic } from './FundingRateGraphic';
 import TracerLoading from '@components/TracerLoading';
 import PoolHealth from '@components/Insurance/PoolHealth';
 
@@ -14,7 +15,7 @@ export const MarketInfo: React.FC<MIProps> = ({ fundingRate, health }: MIProps) 
     ) : (
         <div className="px-5 border-t-2 h-full border-gray-100">
             <Section label={'Funding Rate'}>
-                <FundingRate rate={fundingRate} />
+                <FundingRateGraphic rate={fundingRate} />
             </Section>
             <Section label={'Insurance Pool Health'}>
                 <PoolHealth health={health} />
