+++ components/SummaryInfo/index.ts
@@ -1,6 +1,4 @@
-export { OrderInfo } from './OrderInfo';
 export { AccountMarginInfo } from './AccountMarginInfo';
 export { PnLInfo } from './PnLInfo';
 export { FundingRateGraphic } from './FundingRateGraphic';
 export { MarketInfo } from './MarketInfo';
-export * from './SummaryInfo';
