+++ components/Trade/Advanced/RightPanel/InsuranceInfo.tsx
@@ -1,8 +1,8 @@
-import { InsuranceContext } from '@components/context';
-import { toApproxCurrency } from '@components/libs/utils';
+import { InsuranceContext } from 'context';
+import { toApproxCurrency } from 'libs/utils';
 import React, { useContext } from 'react';
 import styled from 'styled-components';
-import { Section } from './';
+import { Section } from '@components/General';
 import PoolHealth from '@components/Insurance/PoolHealth';
 
 export const InsuranceInfo: React.FC = styled(({ className }) => {
